requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
inbox.py==0.0.6
Logbook==1.5.3
requests==2.25.0
urllib3==1.26.2
